Adds checksum support to Code39ext and Code93ext barcode validators
Extracts a trait from each of Code39 and Code93 for the purposes of
validating the checksum, and then updates the `ext` variants to allow
checksum verification.
